major changes visualization and demo

+++ main.py
@@ -9,6 +9,7 @@ import os
 #import hdf5storage
 from scipy import io
 from sklearn.preprocessing import StandardScaler
+from visualization_paper import results_HCP
 
 #Settings
 def get_args():
@@ -31,7 +32,7 @@ def get_args():
                         help='number of random initializations')
     
     #Preprocessing and training
-    parser.add_argument('--standardise', type=bool, default=True, 
+    parser.add_argument('--standardise', type=bool, default=False, 
                         help='Standardise the data') 
     parser.add_argument('--prediction', type=bool, default=False, 
                         help='Create Train and test sets')
@@ -41,9 +42,9 @@ def get_args():
     #Mising data
     parser.add_argument('--remove', type=bool, default=True,
                         help='Remove data')
-    parser.add_argument('--perc_miss', type=int, default=20,
+    parser.add_argument('--perc_miss', type=int, default=30,
                         help='Percentage of missing data')
-    parser.add_argument('--type_miss', type=str, default='random',
+    parser.add_argument('--type_miss', type=str, default='rows',
                         help='Type of missing data')
     parser.add_argument('--view_miss', type=str, default='2',
                         help='View with missing data')                                            
@@ -62,7 +63,8 @@ else:
     split_data = 'all'              
 
 #Creating path
-res_dir = f'{FLAGS.dir}/{FLAGS.data}/{FLAGS.type}/{FLAGS.method}_{FLAGS.noise}/{FLAGS.m}models/{scenario}/{split_data}/'
+data_dir = f'{FLAGS.dir}/{FLAGS.data}/{FLAGS.type}'
+res_dir = f'{data_dir}/{FLAGS.method}_{FLAGS.noise}/{FLAGS.m}models/{scenario}/{split_data}/'
 if not os.path.exists(res_dir):
         os.makedirs(res_dir)
         
@@ -84,12 +86,10 @@ elif 'ADNI_lowD' in FLAGS.data:
     clinical_data = io.loadmat(f'{data_dir}/Y_splitgender.mat')
 elif 'NSPN' in FLAGS.data:
     data_dir = f'{FLAGS.dir}/{FLAGS.data}/{FLAGS.type}/data'
-    standardise = False
     brain_data = io.loadmat(f'{data_dir}/Xp.mat') 
     clinical_data = io.loadmat(f'{data_dir}/Yp.mat')
 elif 'hcp' in FLAGS.dir:
     data_dir = f'{FLAGS.dir}/{FLAGS.data}/{FLAGS.type}'
-    standardise = False
     brain_data = io.loadmat(f'{data_dir}/X.mat') 
     clinical_data = io.loadmat(f'{data_dir}/Y.mat')               
 
@@ -97,7 +97,7 @@ elif 'hcp' in FLAGS.dir:
 X = [[] for _ in range(2)]
 X[0] = brain_data['X']
 X[1] = clinical_data['Y']
-if standardise is True:
+if FLAGS.standardise is True:
     X[0] = StandardScaler().fit_transform(X[0])
     X[1] = StandardScaler().fit_transform(X[1])
 
@@ -116,46 +116,49 @@ if FLAGS.prediction is True:
 else: 
     X_train = X            
 
-GFAmodel = [[] for _ in range(FLAGS.n_init)]
-for init in range(0, FLAGS.n_init):
-    print("Run:", init+1) 
+#Run model
+filepath = f'{res_dir}GFA_results.dictionary'
+if not os.path.exists(filepath):
+    GFAmodel = [[] for _ in range(FLAGS.n_init)]
+    for init in range(0, FLAGS.n_init):
+        print("Run:", init+1) 
 
-    time_start = time.process_time()
-    d = np.array([X_train[0].shape[1], X_train[1].shape[1]])
-    if FLAGS.remove is True:
-        if 'random' in FLAGS.type_miss:
-            if '1' in FLAGS.view_miss:
-                missing =  np.random.choice([0, 1], size=(X_train[0].shape[0],d[0]), 
-                                        p=[1-FLAGS.perc_miss/100, FLAGS.perc_miss/100])
-                X_train[0][missing == 1] = 'NaN'
-            elif '2' in FLAGS.view_miss:
-                missing =  np.random.choice([0, 1], size=(X_train[1].shape[0],d[1]), 
-                                        p=[1-FLAGS.perc_miss/100, FLAGS.perc_miss/100])
-                X_train[1][missing == 1] = 'NaN' 
+        time_start = time.process_time()
+        d = np.array([X_train[0].shape[1], X_train[1].shape[1]])
+        if FLAGS.remove is True:
+            if 'random' in FLAGS.type_miss:
+                if '1' in FLAGS.view_miss:
+                    missing =  np.random.choice([0, 1], size=(X_train[0].shape[0],d[0]), 
+                                            p=[1-FLAGS.perc_miss/100, FLAGS.perc_miss/100])
+                    X_train[0][missing == 1] = 'NaN'
+                elif '2' in FLAGS.view_miss:
+                    missing =  np.random.choice([0, 1], size=(X_train[1].shape[0],d[1]), 
+                                            p=[1-FLAGS.perc_miss/100, FLAGS.perc_miss/100])
+                    X_train[1][missing == 1] = 'NaN' 
+            
+            elif 'rows' in FLAGS.type_miss:
+                n_rows = int(FLAGS.perc_miss/100 * X[0].shape[0])
+                samples = np.arange(X[0].shape[0])
+                np.random.shuffle(samples)
+                if '1' in FLAGS.view_miss:
+                    X_train[0][samples[0:n_rows],:] = 'NaN'
+                elif '2' in FLAGS.view_miss:
+                    X_train[1][samples[0:n_rows],:] = 'NaN'       
+            
+            GFAmodel[init] = GFAmissing(X_train, FLAGS.m, d)
+        elif 'FA' is FLAGS.noise:   
+            GFAmodel[init] = GFAmissing(X_train, FLAGS.m, d)
+        else:
+            GFAmodel[init] = GFAcomplete(X_train, FLAGS.m, d)
         
-        elif 'rows' in FLAGS.type_miss:
-            n_rows = int(FLAGS.perc_miss/100 * X[0].shape[0])
-            samples = np.arange(X[0].shape[0])
-            np.random.shuffle(samples)
-            if '1' in FLAGS.view_miss:
-                X_train[0][samples[0:n_rows],:] = 'NaN'
-            elif '2' in FLAGS.view_miss:
-                X_train[1][samples[0:n_rows],:] = 'NaN'       
+        if FLAGS.prediction is True:
+            GFAmodel[init].X_test = X_test        
         
-        GFAmodel[init] = GFAmissing(X_train, FLAGS.m, d)
-    elif 'FA' is FLAGS.noise:   
-        GFAmodel[init] = GFAmissing(X_train, FLAGS.m, d)
-    else:
-        GFAmodel[init] = GFAcomplete(X_train, FLAGS.m, d)
-    
-    if FLAGS.prediction is True:
-        GFAmodel[init].X_test = X_test        
-    
-    L = GFAmodel[init].fit(X_train)
-    GFAmodel[init].L = L
-    GFAmodel[init].time_elapsed = (time.process_time() - time_start) 
+        L = GFAmodel[init].fit(X_train)
+        GFAmodel[init].L = L
+        GFAmodel[init].time_elapsed = (time.process_time() - time_start) 
 
-filepath = f'{res_dir}{FLAGS.method}_results.dictionary'
-with open(filepath, 'wb') as parameters:
-
-    pickle.dump(GFAmodel, parameters)
+    with open(filepath, 'wb') as parameters:
+        pickle.dump(GFAmodel, parameters)
+#visualization
+results_HCP(res_dir, data_dir)
